♻️ Refactor: admin api 리팩토링 #9

+++ src/backend/services/admin.service.js
@@ -9,49 +9,91 @@ class AdminService {
   userRepository = new AdminRepository(Users);
   productRepository = new AdminRepository(Products);
 
+  /**
+    * 회원삭제(service)
+    */
   deleteUser = async (userId) => {
     try {
-      // delete 로직
+      const existUser = await this.userRepository.findUser(userId)
+      if (existUser) {
+        return { code: 409, errorMessage: '이미 삭제한 회원입니다.' }
+      }
+
       const deleteUser = await this.userRepository.deleteUser(userId);
       return deleteUser;
     } catch (error) {
-      console.log(error);
+      return { errorMessage: error }
     }
   };
 
-  updateProduct = async (
-    productId,
-    name,
-    price,
-    stock,
-    img_path,
-    description
-  ) => {
+  /**
+    * 상품수정(service)
+    */
+  updateProduct = async ({ productId, name, price, stock, productImage, description }) => {
     try {
-      // 상품 수정
-      const updateProduct = await this.productRepository.updateProduct(
+      /*-----존재하는 상품인지 확인-----*/
+      const product = await this.productRepository.findProduct({ productId })
+      if (product?.destroyTime) {
+        return { code: 412, errorMessage: '이미 삭제된 상품입니다.' }
+      }
+      if (product === null) {
+        return { code: 412, errorMessage: '존재하지 않는 상품입니다.' }
+      }
+
+      /* -----이미지 파일 관련 예외처리----- */
+      if (!productImage) {
+        return {
+          code: 412,
+          errorMessage: '사진을 등록해주세요!'
+        }
+      }
+      const lastDot = productImage.lastIndexOf('.');
+      const ext = productImage.substring(lastDot, productImage.length);
+      if (!ext.match(/\.(jpg|jpeg|png|gif)$/)) {
+        return {
+          code: 412,
+          errorMessage: '이미지 파일만 등록가능합니다.'
+        }
+      }
+      /* -----기타 예외처리-----*/
+      if (!name || !price || !stock || !description) {
+        return {
+          code: 412,
+          errorMessage: '빈칸을 채워주세요!'
+        }
+      }
+      /*-------------------------------*/
+      const updateProduct = await this.productRepository.updateProduct({
         productId,
         name,
         price,
         stock,
-        img_path,
+        productImage,
         description
-      );
+      });
       return updateProduct;
     } catch (error) {
-      console.log(error);
+      return { errorMessage: error }
     }
   };
 
+  /**
+    * 상품삭제(service)
+    */
   deleteProduct = async (productId) => {
     try {
-      // 상품 삭제
-      const deleteProduct = await this.productRepository.deleteProduct(
-        productId
-      );
+      const product = await this.productRepository.findProduct({ productId })
+      if (product?.destroyTime) {
+        return { code: 412, errorMessage: '이미 삭제된 상품입니다.' }
+      }
+      if (product === null) {
+        return { code: 412, errorMessage: '존재하지 않는 상품입니다.' }
+      }
+
+      const deleteProduct = await this.productRepository.deleteProduct({ productId });
       return deleteProduct;
     } catch (error) {
-      console.log(error);
+      return { errorMessage: error }
     }
   };
